refactor: 코드 리뷰 반영
- 상수 선언
- 불필요한 생성자 제거
- View Class 각각 input, result 분리
- 입력인자 변경
