+++ src/main/java/step3/RacingGame.java
@@ -2,17 +2,26 @@ package step3;
 
 import step3.domain.Cars;
 import step3.service.RandomRacingRule;
-import step3.view.RacingGameView;
+import step3.view.RacingGameOutputView;
+
+import java.util.Arrays;
+import java.util.List;
 
 public class RacingGame {
 
+  private static final String SPLIT_FLAG = ",";
+
   public void start(String carNames, int gameCount) {
-    Cars cars = new Cars(carNames);
+    Cars cars = new Cars(splitCarNames(carNames));
     for (int i = 0; i < gameCount; i++) {
       cars.race(new RandomRacingRule());
-      RacingGameView.printRacingGameResult(cars);
+      RacingGameOutputView.printRacingGameResult(cars);
     }
-    RacingGameView.printWinnerResult(cars);
+    RacingGameOutputView.printWinnerResult(cars);
+  }
+
+  private List<String> splitCarNames(String carNames) {
+    return Arrays.asList(carNames.split(SPLIT_FLAG));
   }
 
 }
