+++ src/main/java/step3/RacingGameMain.java
@@ -1,11 +1,11 @@
 package step3;
 
-import step3.view.RacingGameView;
+import step3.view.RacingGameInputView;
 
 public class RacingGameMain {
   public static void main(String[] args) {
-    String carNames = RacingGameView.getCarNamesFromUser();
-    int gameCount = RacingGameView.getGameCountFromUser();
+    String carNames = RacingGameInputView.getCarNamesFromUser();
+    int gameCount = RacingGameInputView.getGameCountFromUser();
 
     System.out.println("실행결과");
     RacingGame game = new RacingGame();
